requirements.txt
loguru==0.6.0
pydantic==1.10.2
sqlmodel==0.0.8
psycopg2-binary==2.9.5
requests==2.28.1
python-telegram-bot==20.b0
BeautifulSoup4==4.11.1
openai==0.25.0
readability-lxml==0.8.1
transformers==4.25.1
markdown==3.4.1
pdfminer.six==20221105
openai==0.25.0
readability-lxml==0.8.1
pdfminer.six==20221105
sentence_transformers==2.2.2










